js/xpconnect/tests/mochitest/test_bug628410.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=628410
-->
<head>
  <title>Test for Bug 628410</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=628410">Mozilla Bug 628410</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>

<pre id="test">
<script type="application/javascript">

window.toString();

if (SpecialPowers.Services.prefs.getBoolPref("extensions.InstallTrigger.enabled") &&
    SpecialPowers.Services.prefs.getBoolPref("extensions.InstallTriggerImpl.enabled")) {
  InstallTrigger + "";
}

console + "";
ok(true, "Things didn't throw");

</script>
</pre>
</body>
</html>
